use TextImproved native component on Android
